Create new 'widgets' directory and move related files into it
