+++ app/src/main/kotlin/uk/gov/govuk/ui/widgets/AlertBanner.kt
@@ -1,4 +1,4 @@
-package uk.gov.govuk.alertbanner.ui
+package uk.gov.govuk.ui.widgets
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
