+++ app/src/main/kotlin/uk/gov/govuk/ui/widgets/HomeWidgets.kt
@@ -1,16 +1,14 @@
-package uk.gov.govuk.ui
+package uk.gov.govuk.ui.widgets
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
-import uk.gov.govuk.alertbanner.ui.AlertBanner
 import uk.gov.govuk.design.ui.component.LargeVerticalSpacer
 import uk.gov.govuk.topics.navigation.navigateToTopic
 import uk.gov.govuk.topics.navigation.navigateToTopicsAll
 import uk.gov.govuk.topics.navigation.navigateToTopicsEdit
 import uk.gov.govuk.topics.ui.widget.TopicsWidget
 import uk.gov.govuk.ui.model.HomeWidget
-import uk.gov.govuk.userfeedbackbanner.ui.UserFeedbackBanner
 import uk.gov.govuk.visited.navigation.VISITED_GRAPH_ROUTE
 import uk.gov.govuk.visited.ui.widget.VisitedWidget
 import uk.govuk.app.local.navigation.LOCAL_GRAPH_ROUTE
